Switch params for HTTP requests based on Rails version
Rails 5 deprecated sequential arguments in favor of keyword arguments
(params:, headers:).  Since we want to support multiple versions of
Rails without showing deprecation warnings (or totally failing in Rails
5.1), use keyword arguments when working with newer versions of Rails.

The strategy keeps things as close as possible to the Rails 5 version in
Apivore::Validator to create minimal work later to remove the shim when
Rails 5 becomes a gem dependency.

Because Ruby 2.0 (which we still support) doesn't allow required keyword
arguments, the shim uses defaults, but theoretically the defaulting
should be handled in Apivore::Validator (as it currently does).